+++ middleware/index.js
@@ -1,0 +1,8 @@
+import validateId from "./validationsContacts/validateId";
+import validateCreate from "./validationsContacts/validateCreate";
+import validateUpdate from "./validationsContacts/validateUpdate";
+export default {
+  validateCreate,
+  validateUpdate,
+  validateId,
+};
