+++ middleware/validationsContacts/validateCreate.js
@@ -1,0 +1,17 @@
+import Joi from "joi";
+
+const createSchema = Joi.object({
+  name: Joi.string().required(),
+  email: Joi.string().email().required(),
+  phone: Joi.string().required(),
+});
+
+const validateCreate = async (req, res, next) => {
+  try {
+    await createSchema.validateAsync(req.body);
+  } catch (error) {
+    return res.status(400).json({ message: error.message.replace(/"/g, "") });
+  }
+  next();
+};
+export default validateCreate;
